No JIRA: In server deployment mode, classNotFoundException for net.hedtech.banner.testing.AcademicYearForTesting:
Reason: This is a test class. All test packages are deleted from WAR through eventCreateWarStart closure of _Events.groovy of Application codebase. So, these test classes had to be moved into a separate test hibernate config file to be loaded only from plugin and not to be loaded from application at all to not searching during WAR mode.